requirements.txt
beautifulsoup4==4.9.3
Keras==2.4.3
newspaper3k==0.2.8
nltk==3.5
numpy==1.18.5
pandas==1.1.4
tensorflow==2.3.1
textblob==0.15.3
